requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
cohere==4.37
gunicorn==21.2.0